Aug-28-2024: separate code into components
